* Minnor improvements
